+++ .gitignore
@@ -5,4 +5,5 @@ vendor/
 reports/
 src/csv/
 src/log/
+src/reports/
 src/test.py
