+++ src/log/coeus.log
@@ -214,3 +214,31 @@
 [2023-02-21 21:09:49,799-INFO-getAllFilesFromRepository-38]: Searching all relevant files
 [2023-02-21 21:10:01,273-INFO-getAllFilesFromRepository-38]: Searching all relevant files
 [2023-02-21 21:10:17,937-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-23 14:03:20,891-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-23 14:35:23,338-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-23 14:35:23,361-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-23 14:41:59,665-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-23 14:41:59,685-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-23 14:42:34,467-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-23 14:55:52,502-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 17:59:40,683-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 17:59:50,688-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 18:03:03,574-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 18:04:35,444-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 18:04:45,053-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 18:05:12,458-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 18:17:06,964-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 18:17:23,411-INFO-getAllFilesFromRepository-38]: Searching all relevant files
+[2023-02-26 21:25:47,673-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 21:26:02,689-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 21:37:16,837-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 21:37:53,656-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 21:38:20,672-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 21:40:26,787-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 22:07:54,309-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 22:09:28,162-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 22:13:47,336-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 22:14:02,596-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 22:14:31,512-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 22:14:48,778-INFO-getAllFilesFromRepository-40]: Searching all relevant files
+[2023-02-26 22:15:31,210-INFO-getAllFilesFromRepository-40]: Searching all relevant files
