+++ tests/context.py
@@ -1,0 +1,5 @@
+import sys
+import os
+sys.path.append(os.path.abspath(os.path.join('..')))
+
+from src import *
